Revert "Created dedicated print service"
This reverts commit 6b8e46497718acb4da909790c9528ec9c9cc71a0.
