+++ src-ui/src/app/components/document-detail/document-detail.component.ts
@@ -69,7 +69,6 @@ import {
   PermissionsService,
   PermissionType,
 } from 'src/app/services/permissions.service'
-import { PrintService } from 'src/app/services/print.service'
 import { CorrespondentService } from 'src/app/services/rest/correspondent.service'
 import { CustomFieldsService } from 'src/app/services/rest/custom-fields.service'
 import { DocumentTypeService } from 'src/app/services/rest/document-type.service'
@@ -180,7 +179,6 @@ export class DocumentDetailComponent
   implements OnInit, OnDestroy, DirtyComponent
 {
   private documentsService = inject(DocumentService)
-  private printService = inject(PrintService)
   private route = inject(ActivatedRoute)
   private correspondentService = inject(CorrespondentService)
   private documentTypeService = inject(DocumentTypeService)
@@ -1402,7 +1400,40 @@ export class DocumentDetailComponent
       this.document.id,
       false
     )
-    this.printService.printDocument(printUrl)
+    this.http.get(printUrl, { responseType: 'blob' }).subscribe({
+      next: (blob) => {
+        if (this.deviceDetectorService.isMobile()) {
+          const blobUrl = URL.createObjectURL(blob)
+          window.open(blobUrl, '_blank')
+          setTimeout(() => URL.revokeObjectURL(blobUrl), 1000)
+          return
+        }
+        const blobUrl = URL.createObjectURL(blob)
+        const iframe = document.createElement('iframe')
+        iframe.style.display = 'none'
+        iframe.src = blobUrl
+        document.body.appendChild(iframe)
+        iframe.onload = () => {
+          try {
+            iframe.contentWindow.focus()
+            iframe.contentWindow.print()
+            iframe.contentWindow.onafterprint = () => {
+              document.body.removeChild(iframe)
+              URL.revokeObjectURL(blobUrl)
+            }
+          } catch (err) {
+            this.toastService.showError($localize`Print failed.`, err)
+            document.body.removeChild(iframe)
+            URL.revokeObjectURL(blobUrl)
+          }
+        }
+      },
+      error: () => {
+        this.toastService.showError(
+          $localize`Error loading document for printing.`
+        )
+      },
+    })
   }
 
   public openShareLinks() {
