+++ src-ui/src/app/components/document-list/document-card-large/document-card-large.component.html
@@ -77,7 +77,7 @@
               <a class="btn btn-sm btn-outline-secondary" [href]="getDownloadUrl()">
                 <i-bs name="download"></i-bs>&nbsp;<span class="d-none d-md-inline" i18n>Download</span>
               </a>
-              <a class="btn btn-sm btn-outline-secondary" (click)="printDocument()">
+              <a class="btn btn-sm btn-outline-secondary" [href]="getDownloadUrl()">
                 <i-bs name="printer"></i-bs>&nbsp;<span class="d-none d-md-inline" i18n>Print</span>
               </a>
             } @else {
