+++ src-ui/src/app/components/document-list/document-card-small/document-card-small.component.html
@@ -149,7 +149,7 @@
             <a [href]="getDownloadUrl()" class="btn btn-sm btn-outline-secondary" title="Download" i18n-title (click)="$event.stopPropagation()">
               <i-bs name="download"></i-bs>
             </a>
-            <a (click)="printDocument()" class="btn btn-sm btn-outline-secondary" title="Print" i18n-title>
+            <a [href]="getDownloadUrl()" class="btn btn-sm btn-outline-secondary" title="Print" i18n-title>
               <i-bs name="printer"></i-bs>
             </a>
           } @else {
