+++ src-ui/src/app/components/document-list/document-card-small/document-card-small.component.ts
@@ -30,7 +30,6 @@ import { DocumentTypeNamePipe } from 'src/app/pipes/document-type-name.pipe'
 import { IsNumberPipe } from 'src/app/pipes/is-number.pipe'
 import { StoragePathNamePipe } from 'src/app/pipes/storage-path-name.pipe'
 import { UsernamePipe } from 'src/app/pipes/username.pipe'
-import { PrintService } from 'src/app/services/print.service'
 import { DocumentService } from 'src/app/services/rest/document.service'
 import { SettingsService } from 'src/app/services/settings.service'
 import { CustomFieldDisplayComponent } from '../../common/custom-field-display/custom-field-display.component'
@@ -66,7 +65,6 @@ export class DocumentCardSmallComponent
   implements AfterViewInit
 {
   private documentService = inject(DocumentService)
-  private printService = inject(PrintService)
   settingsService = inject(SettingsService)
 
   DisplayField = DisplayField
@@ -120,14 +118,6 @@ export class DocumentCardSmallComponent
 
   getDownloadUrl() {
     return this.documentService.getDownloadUrl(this.document.id)
-  }
-
-  printDocument() {
-    const printUrl = this.documentService.getDownloadUrl(
-      this.document.id,
-      false
-    )
-    this.printService.printDocument(printUrl)
   }
 
   get tagIDs() {
